Add group integration test
Split index.ts into app.ts so I can export the Express 'app' for use in tests as well as pass it to the server in index.ts.

Add unique constraint on group name and organization_id so that duplicate groups are not added.

Move db schema statements to schema.ts so we can reuse them in the test db setup.

format cookie-helper